Allow require("store/plugins/X") and require("store/storages/X") by moving src/plugin to plugins and src/storage storages; Rename builds to dist
