+++ plugins/all.js
