+++ plugins/all_tests.js
