+++ plugins/lib/json2.js
