+++ src/builds/cross-browser.js
@@ -1,8 +1,8 @@
 var engine = require('../engine')
 
-var allStorages = require('../storage/all')
+var allStorages = require('../../storages/all')
 var plugins = [
-	require('../plugin/json2'),
+	require('../../plugins/json2'),
 ]
 
 module.exports = engine.createStore(allStorages, plugins)
